feat: Implement chat functionality with PortfolioChat, ChatInput, ChatMessage, and SuggestedQuestions components
