--- client/src/components/PortfolioChat.tsx
@@ -1,549 +1,0 @@
-import { Button } from "@/components/ui/button";
-import { Card } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { useToast } from "@/hooks/use-toast";
-import { apiRequest } from "@/lib/queryClient";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { Send, Bot, User, Plus } from "lucide-react";
-import { useState, useEffect, useMemo } from "react";
-import ReactMarkdown from "react-markdown";
-import remarkGfm from "remark-gfm";
-
-interface Message {
-  id: string;
-  content: string;
-  sender: "user" | "ai";
-  timestamp: Date;
-}
-
-interface PortfolioChatProps {
-  onSendMessage?: (message: string) => void;
-  portfolio?: any; // Add portfolio prop
-}
-
-export default function PortfolioChat({
-  onSendMessage,
-  portfolio,
-}: PortfolioChatProps) {
-  const [message, setMessage] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const [isNewChat, setIsNewChat] = useState(false);
-  const [showNewChatSuccess, setShowNewChatSuccess] = useState(false);
-  const [streamingMessage, setStreamingMessage] = useState<string>("");
-  const [isStreaming, setIsStreaming] = useState(false);
-  const { toast } = useToast();
-
-  const queryClient = useQueryClient();
-  const [portfolioId, setPortfolioId] = useState<string | null>(null);
-  // Must be declared before any early returns to keep hook order stable
-  const [optimisticMessages, setOptimisticMessages] = useState<Message[]>([]);
-
-  // Fetch portfolio on mount
-  const { data: portfolioData } = useQuery<{
-    id: string;
-    allocations: any[];
-    totalValue: number;
-    totalReturn: number;
-  }>({
-    queryKey: ["/api/portfolio"],
-  });
-
-  useEffect(() => {
-    if (portfolioData?.id) {
-      setPortfolioId(portfolioData.id);
-    }
-  }, [portfolioData]);
-
-  // Hide the new chat success message after 3 seconds
-  useEffect(() => {
-    if (showNewChatSuccess) {
-      const timer = setTimeout(() => {
-        setShowNewChatSuccess(false);
-      }, 3000);
-      return () => clearTimeout(timer);
-    }
-  }, [showNewChatSuccess]);
-
-  const { data: messagesData = [] } = useQuery({
-    queryKey: ["/api/portfolio", portfolioId, "messages"],
-    queryFn: async () => {
-      const res = await apiRequest(
-        "GET",
-        `/api/portfolio/${portfolioId}/messages`,
-      );
-      const data = await res.json();
-      return data.map((msg: any) => ({
-        id: msg.id,
-        content: msg.content,
-        sender: msg.sender,
-        timestamp: new Date(msg.createdAt),
-      })) as Message[];
-    },
-    enabled: !!portfolioId,
-    staleTime: 2 * 60 * 1000, // 2 minutes - messages can change but don't need to refetch constantly
-    gcTime: 15 * 60 * 1000, // 15 minutes
-  });
-
-  const sendMessageWithStreaming = async (content: string) => {
-    try {
-      const response = await fetch(`/api/portfolio/${portfolioId}/messages`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        credentials: "include",
-        body: JSON.stringify({ content }),
-      });
-
-      if (!response.ok) {
-        throw new Error("Failed to send message");
-      }
-
-      const reader = response.body?.getReader();
-      const decoder = new TextDecoder();
-
-      if (!reader) {
-        throw new Error("No reader available");
-      }
-
-      let userMessageData: any = null;
-      let accumulatedContent = "";
-      setIsStreaming(true);
-
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-
-        const chunk = decoder.decode(value);
-        const lines = chunk.split("\n");
-
-        for (const line of lines) {
-          if (line.startsWith("data: ")) {
-            try {
-              const data = JSON.parse(line.slice(6));
-
-              if (data.type === "userMessage") {
-                userMessageData = data.data;
-              } else if (data.type === "chunk") {
-                accumulatedContent += data.data;
-                setStreamingMessage(accumulatedContent);
-              } else if (data.type === "complete") {
-                // Message complete, refresh the list
-                setIsStreaming(false);
-                setStreamingMessage("");
-                if (!isNewChat) {
-                  queryClient.invalidateQueries({
-                    queryKey: ["messages", portfolioId],
-                  });
-                } else {
-                  const currentMessages =
-                    (queryClient.getQueryData(["messages", portfolioId]) as
-                      | Message[]
-                      | undefined) || [];
-                  const transformedMessages = [
-                    {
-                      id: userMessageData.id,
-                      content: userMessageData.content,
-                      sender: userMessageData.sender,
-                      timestamp: new Date(userMessageData.createdAt),
-                    },
-                    {
-                      id: data.data.id,
-                      content: data.data.content,
-                      sender: data.data.sender,
-                      timestamp: new Date(data.data.createdAt),
-                    },
-                  ] as Message[];
-                  queryClient.setQueryData(
-                    ["messages", portfolioId],
-                    [...currentMessages, ...transformedMessages],
-                  );
-                  setIsNewChat(false);
-                }
-              } else if (data.type === "error") {
-                throw new Error(data.data.message);
-              }
-            } catch (e) {
-              console.error("Error parsing SSE data:", e);
-            }
-          }
-        }
-      }
-    } catch (error) {
-      console.error("Streaming error:", error);
-      setIsStreaming(false);
-      setStreamingMessage("");
-      toast({
-        title: "Error",
-        description: "Failed to send message. Please try again.",
-        variant: "destructive",
-      });
-      throw error;
-    }
-  };
-
-  // Compute suggested questions early to keep hook order stable
-  const suggestedQuestions = useMemo(() => {
-    try {
-      if (
-        !portfolio?.allocations ||
-        !Array.isArray(portfolio.allocations) ||
-        portfolio.allocations.length === 0
-      ) {
-        return ["What should I invest in as a beginner?"];
-      }
-
-      const questions = [
-        "Explain my current allocation",
-        "What if the market drops 20%?",
-        "How can I reduce risk?",
-        "Should I rebalance?",
-        "What about ESG investing?",
-        "How much should I invest monthly?",
-        "When should I retire?",
-        "What about international stocks?",
-        "How to handle market volatility?",
-        "What are the tax implications?",
-      ];
-
-      const bondPct = portfolio.allocations.reduce(
-        (sum: number, a: any) =>
-          sum + (a?.assetType === "Bonds" ? a?.percentage || 0 : 0),
-        0,
-      );
-      const stockPct = portfolio.allocations.reduce(
-        (sum: number, a: any) =>
-          sum + (a?.assetType?.includes("Equity") ? a?.percentage || 0 : 0),
-        0,
-      );
-      const esg = portfolio.allocations.some((a: any) =>
-        a?.name?.includes("ESG"),
-      );
-
-      // Add personalized questions based on portfolio
-      if (bondPct > 50) {
-        questions.push("Why is my portfolio so conservative?");
-      }
-      if (stockPct > 70) {
-        questions.push("How can I reduce risk in my portfolio?");
-      }
-      if (esg) {
-        questions.push("Tell me more about the ESG focus in my investments");
-      } else {
-        questions.push("Should I consider ESG investments?");
-      }
-
-      return questions.slice(0, 8); // Show more questions for better scrolling
-    } catch (error) {
-      console.error("Error generating suggested questions:", error);
-      return ["What should I invest in as a beginner?"];
-    }
-  }, [portfolio]);
-
-  // If !portfolioId return loading
-  if (!portfolioId) {
-    if (portfolioData === null) {
-      return (
-        <Card className="flex h-[600px] flex-col">
-          <div className="border-b p-4">
-            <h3 className="flex items-center gap-2 font-semibold">
-              <Bot className="h-5 w-5 text-primary" />
-              Chat with Your Portfolio AI
-            </h3>
-            <p className="text-sm text-muted-foreground">
-              No portfolio generated yet. Please complete your investor profile
-              and generate a portfolio to start chatting.
-            </p>
-          </div>
-          <div className="flex flex-1 items-center justify-center p-4">
-            <div className="text-center">
-              <Bot className="mx-auto mb-4 h-12 w-12 text-muted-foreground" />
-              <p className="text-muted-foreground">
-                Portfolio AI is ready when you are.
-              </p>
-            </div>
-          </div>
-        </Card>
-      );
-    }
-    return <div>Loading portfolio...</div>;
-  }
-
-  const messages = [...messagesData, ...optimisticMessages];
-
-  const handleNewChat = async () => {
-    if (!portfolioId) return;
-
-    try {
-      // Delete all messages from the database
-      await apiRequest("DELETE", `/api/portfolio/${portfolioId}/messages`);
-
-      // Clear optimistic messages and input
-      setOptimisticMessages([]);
-      setMessage("");
-
-      // Set the messages query data to empty array immediately
-      queryClient.setQueryData(["messages", portfolioId], []);
-
-      // Mark that we're in new chat mode to prevent query invalidation on first message
-      setIsNewChat(true);
-
-      // Show inline success message near the button
-      setShowNewChatSuccess(true);
-    } catch (error) {
-      console.error("Error starting new chat:", error);
-      toast({
-        title: "Error",
-        description: "Failed to start new chat. Please try again.",
-        variant: "destructive",
-      });
-    }
-  };
-
-  const handleSendMessage = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!message.trim() || isLoading || !portfolioId) return;
-
-    // Optimistically add user message immediately
-    const tempId = `temp-${Date.now()}`;
-    const optimistic: Message = {
-      id: tempId,
-      content: message,
-      sender: "user",
-      timestamp: new Date(),
-    };
-    setOptimisticMessages((prev) => [...prev, optimistic]);
-    setIsLoading(true);
-
-    const currentMessage = message;
-    setMessage("");
-
-    try {
-      await sendMessageWithStreaming(currentMessage);
-      setIsLoading(false);
-      setOptimisticMessages([]);
-    } catch (error) {
-      setIsLoading(false);
-      setOptimisticMessages((prev) => prev.filter((m) => m.id !== tempId));
-      setMessage(currentMessage); // Restore message on error
-    }
-  };
-
-  const formatTime = (date: Date) => {
-    return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
-  };
-
-  return (
-    <Card className="flex h-[600px] w-full min-w-0 max-w-full flex-col">
-      <div className="border-b p-4">
-        <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between sm:gap-0">
-          <div className="min-w-0 flex-1">
-            <h3 className="flex items-center gap-2 text-base font-semibold sm:text-lg">
-              <Bot className="h-5 w-5 flex-shrink-0 text-primary" />
-              <span className="truncate">Chat with Your Portfolio AI</span>
-            </h3>
-            <p className="mt-1 text-sm text-muted-foreground">
-              Ask questions about your investments, market trends, or portfolio
-              strategy
-            </p>
-          </div>
-          <div className="flex flex-col items-start gap-2 sm:flex-shrink-0 sm:items-end">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={handleNewChat}
-              className="flex min-h-[36px] w-full touch-manipulation items-center gap-2 sm:w-auto"
-              data-testid="button-new-chat"
-            >
-              <Plus className="h-4 w-4" />
-              New Chat
-            </Button>
-            {showNewChatSuccess && (
-              <div className="text-xs font-medium text-green-600 duration-300 animate-in fade-in-0 slide-in-from-top-1 dark:text-green-400">
-                ✓ Conversation cleared
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-
-      <ScrollArea className="flex-1 p-4">
-        <div className="space-y-4">
-          {messages.map((msg: Message) => (
-            <div
-              key={msg.id}
-              className={`flex gap-3 ${msg.sender === "user" ? "justify-end" : "justify-start"}`}
-            >
-              {msg.sender === "ai" && (
-                <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-primary text-primary-foreground">
-                  <Bot className="h-4 w-4" />
-                </div>
-              )}
-
-              <div
-                className={`max-w-[85%] rounded-lg p-3 sm:max-w-[80%] ${
-                  msg.sender === "user"
-                    ? "bg-primary text-primary-foreground"
-                    : "bg-muted"
-                }`}
-              >
-                {msg.sender === "ai" ? (
-                  <div className="prose prose-sm dark:prose-invert max-w-none">
-                    <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                      {msg.content}
-                    </ReactMarkdown>
-                  </div>
-                ) : (
-                  <p className="text-sm">{msg.content}</p>
-                )}
-                <p className={`mt-1 text-xs opacity-70`}>
-                  {formatTime(msg.timestamp)}
-                </p>
-              </div>
-
-              {msg.sender === "user" && (
-                <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-secondary text-secondary-foreground">
-                  <User className="h-4 w-4" />
-                </div>
-              )}
-            </div>
-          ))}
-
-          {/* Streaming AI message */}
-          {isStreaming && streamingMessage && (
-            <div className="flex justify-start gap-3 duration-200 animate-in fade-in">
-              <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-primary text-primary-foreground">
-                <Bot className="h-4 w-4" />
-              </div>
-              <div className="max-w-[85%] rounded-lg bg-muted p-3 sm:max-w-[80%]">
-                <div className="prose prose-sm dark:prose-invert max-w-none">
-                  <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                    {streamingMessage}
-                  </ReactMarkdown>
-                </div>
-                <div className="mt-2 flex items-center gap-1">
-                  <div
-                    className="h-1.5 w-1.5 animate-bounce rounded-full bg-primary"
-                    style={{ animationDelay: "0ms" }}
-                  ></div>
-                  <div
-                    className="h-1.5 w-1.5 animate-bounce rounded-full bg-primary"
-                    style={{ animationDelay: "150ms" }}
-                  ></div>
-                  <div
-                    className="h-1.5 w-1.5 animate-bounce rounded-full bg-primary"
-                    style={{ animationDelay: "300ms" }}
-                  ></div>
-                </div>
-              </div>
-            </div>
-          )}
-
-          {isLoading && !isStreaming && (
-            <div className="flex gap-3">
-              <div className="flex h-8 w-8 items-center justify-center rounded-full bg-primary text-primary-foreground">
-                <Bot className="h-4 w-4" />
-              </div>
-              <div className="rounded-lg bg-muted p-3">
-                <div className="flex gap-1">
-                  <div className="h-2 w-2 animate-bounce rounded-full bg-muted-foreground"></div>
-                  <div
-                    className="h-2 w-2 animate-bounce rounded-full bg-muted-foreground"
-                    style={{ animationDelay: "0.1s" }}
-                  ></div>
-                  <div
-                    className="h-2 w-2 animate-bounce rounded-full bg-muted-foreground"
-                    style={{ animationDelay: "0.2s" }}
-                  ></div>
-                </div>
-              </div>
-            </div>
-          )}
-        </div>
-      </ScrollArea>
-
-      <div className="relative overflow-hidden border-t p-4">
-        <div className="mb-3 flex items-center justify-between">
-          <span className="text-sm font-medium text-foreground">
-            Suggested questions
-          </span>
-          <span className="flex items-center gap-1 text-xs text-muted-foreground">
-            <span className="hidden sm:inline">Swipe horizontally</span>
-            <span className="sm:hidden">← Swipe →</span>
-          </span>
-        </div>
-        <div className="relative">
-          <div
-            className="suggested-questions-container scrollbar-hide -mx-4 flex gap-2 overflow-x-auto overflow-y-hidden px-4 pb-2 [&::-webkit-scrollbar]:hidden"
-            style={{
-              WebkitOverflowScrolling: "touch",
-              scrollBehavior: "smooth",
-              overscrollBehaviorX: "contain",
-              scrollbarWidth: "none",
-              msOverflowStyle: "none",
-              touchAction: "pan-x",
-              scrollSnapType: "x mandatory",
-              // Ensure proper scrolling on mobile
-              minHeight: "56px", // Ensure touch target size and better scrolling
-              maxHeight: "120px", // Prevent excessive height on mobile
-            }}
-          >
-            {suggestedQuestions.map((q, i) => (
-              <Button
-                key={i}
-                variant="outline"
-                size="sm"
-                onClick={(e) => {
-                  e.preventDefault();
-                  setMessage(q);
-                  handleSendMessage(e);
-                }}
-                className="scroll-snap-align-start h-auto min-w-fit flex-shrink-0 touch-manipulation whitespace-nowrap rounded-full border-primary/20 px-4 py-3 text-xs transition-all duration-200 hover:border-primary/40 hover:bg-primary/5 active:scale-95 sm:px-5 sm:text-sm"
-                style={{
-                  // Ensure minimum touch target size on mobile
-                  minHeight: "48px",
-                  minWidth: "max-content",
-                  // Prevent text selection during scroll
-                  userSelect: "none",
-                  WebkitUserSelect: "none",
-                  // Better mobile touch scrolling
-                  touchAction: "manipulation",
-                }}
-              >
-                {q}
-              </Button>
-            ))}
-
-            {/* Add padding at the end for better scrolling experience */}
-            <div className="w-1 flex-shrink-0"></div>
-          </div>
-
-          {/* Gradient fades removed to simplify appearance */}
-        </div>
-      </div>
-
-      <form onSubmit={handleSendMessage} className="border-t p-4">
-        <div className="flex gap-3">
-          <Input
-            value={message}
-            onChange={(e) => setMessage(e.target.value)}
-            placeholder="Ask about your portfolio..."
-            disabled={isLoading}
-            className="min-h-[44px] touch-manipulation"
-            data-testid="input-chat-message"
-          />
-          <Button
-            type="submit"
-            size="icon"
-            disabled={!message.trim() || isLoading}
-            className="min-h-[44px] min-w-[44px] touch-manipulation"
-            data-testid="button-send-message"
-          >
-            <Send className="h-4 w-4" />
-          </Button>
-        </div>
-      </form>
-    </Card>
-  );
-}
