+++ client/src/components/chat/ChatMessage.tsx
@@ -1,0 +1,53 @@
+import { Bot, User } from "lucide-react";
+import ReactMarkdown from "react-markdown";
+import remarkGfm from "remark-gfm";
+import type { Message } from "@/hooks/usePortfolioMessages";
+
+interface ChatMessageProps {
+  message: Message;
+}
+
+export function ChatMessage({ message }: ChatMessageProps) {
+  const formatTime = (date: Date) => {
+    return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
+  };
+
+  return (
+    <div
+      className={`flex gap-3 ${message.sender === "user" ? "justify-end" : "justify-start"}`}
+    >
+      {message.sender === "ai" && (
+        <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-primary text-primary-foreground">
+          <Bot className="h-4 w-4" />
+        </div>
+      )}
+
+      <div
+        className={`max-w-[85%] rounded-lg p-3 sm:max-w-[80%] ${
+          message.sender === "user"
+            ? "bg-primary text-primary-foreground"
+            : "bg-muted"
+        }`}
+      >
+        {message.sender === "ai" ? (
+          <div className="prose prose-sm dark:prose-invert max-w-none">
+            <ReactMarkdown remarkPlugins={[remarkGfm]}>
+              {message.content}
+            </ReactMarkdown>
+          </div>
+        ) : (
+          <p className="text-sm">{message.content}</p>
+        )}
+        <p className="mt-1 text-xs opacity-70">
+          {formatTime(message.timestamp)}
+        </p>
+      </div>
+
+      {message.sender === "user" && (
+        <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-secondary text-secondary-foreground">
+          <User className="h-4 w-4" />
+        </div>
+      )}
+    </div>
+  );
+}
