+++ client/src/components/chat/index.ts
@@ -1,0 +1,4 @@
+export { default as PortfolioChat } from "./PortfolioChat";
+export { ChatMessage } from "./ChatMessage";
+export { ChatInput } from "./ChatInput";
+export { SuggestedQuestions } from "./SuggestedQuestions";
